- Small datatype fix to squelch compiler warnings in release
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1608 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/Skills.cpp
@@ -137,7 +137,7 @@ void Skills::useSkill(Player *player, PacketReader &packet) {
 		{
 			uint8_t players = packet.getByte();
 			for (uint8_t i = 0; i < players; i++) {
-				int16_t playerid = packet.getInt();
+				int32_t playerid = packet.getInt();
 				if (Player *target = Players::Instance()->getPlayer(playerid)) {
 					SkillsPacket::showSkill(player, skillid, level, true, true);
 					SkillsPacket::showSkill(player, skillid, level, true);
